requirements.txt
fastapi
uvicorn
pyserial
pandas
numpy
scipy
jinja2
reportlab